--- generate_inkatlas.py
@@ -1,363 +1,0 @@
-#      ▓█████▄  ██▀███      ██▓███   ██▀███  ▓█████   ██████ ▄▄▄█████▓ ▒█████    ██████ 
-#      ▒██▀ ██▌▓██ ▒ ██▒   ▓██░  ██▒▓██ ▒ ██▒▓█   ▀ ▒██    ▒ ▓  ██▒ ▓▒▒██▒  ██▒▒██    ▒ 
-#      ░██   █▌▓██ ░▄█ ▒   ▓██░ ██▓▒▓██ ░▄█ ▒▒███   ░ ▓██▄   ▒ ▓██░ ▒░▒██░  ██▒░ ▓██▄   
-#      ░▓█▄   ▌▒██▀▀█▄     ▒██▄█▓▒ ▒▒██▀▀█▄  ▒▓█  ▄   ▒   ██▒░ ▓██▓ ░ ▒██   ██░  ▒   ██▒
-#      ░▒████▓ ░██▓ ▒██▒   ▒██▒ ░  ░░██▓ ▒██▒░▒████▒▒██████▒▒  ▒██▒ ░ ░ ████▓▒░▒██████▒▒
-#       ▒▒▓  ▒ ░ ▒▓ ░▒▓░   ▒▓▒░ ░  ░░ ▒▓ ░▒▓░░░ ▒░ ░▒ ▒▓▒ ▒ ░  ▒ ░░   ░ ▒░▒░▒░ ▒ ▒▓▒ ▒ ░
-#       ░ ▒  ▒   ░▒ ░ ▒░   ░▒ ░       ░▒ ░ ▒░ ░ ░  ░░ ░▒  ░ ░    ░      ░ ▒ ▒░ ░ ░▒  ░ ░
-#       ░ ░  ░   ░░   ░    ░░         ░░   ░    ░   ░  ░  ░    ░      ░ ░ ░ ▒  ░  ░  ░  
-#         ░       ░                    ░        ░  ░      ░               ░ ░        ░  
-#   ███▄ ▄███▓ ▄▄▄        ▄████  ███▄    █  ██▓  █████▒██▓ ▄████▄  ▓█████  ███▄    █ ▄▄▄█████▓
-#  ▓██▒▀█▀ ██▒▒████▄     ██▒ ▀█▒ ██ ▀█   █ ▓██▒▓██   ▒▓██▒▒██▀ ▀█  ▓█   ▀  ██ ▀█   █ ▓  ██▒ ▓▒
-#  ▓██    ▓██░▒██  ▀█▄  ▒██░▄▄▄░▓██  ▀█ ██▒▒██▒▒████ ░▒██▒▒▓█    ▄ ▒███   ▓██  ▀█ ██▒▒ ▓██░ ▒░
-#  ▒██    ▒██ ░██▄▄▄▄██ ░▓█  ██▓▓██▒  ▐▌██▒░██░░▓█▒  ░░██░▒▓▓▄ ▄██▒▒▓█  ▄ ▓██▒  ▐▌██▒░ ▓██▓ ░ 
-#  ▒██▒   ░██▒ ▓█   ▓██▒░▒▓███▀▒▒██░   ▓██░░██░░▒█░   ░██░▒ ▓███▀ ░░▒████▒▒██░   ▓██░  ▒██▒ ░ 
-#  ░ ▒░   ░  ░ ▒▒   ▓▒█░ ░▒   ▒ ░ ▒░   ▒ ▒ ░▓   ▒ ░   ░▓  ░ ░▒ ▒  ░░░ ▒░ ░░ ▒░   ▒ ▒   ▒ ░░   
-#  ░  ░      ░  ▒   ▒▒ ░  ░   ░ ░ ░░   ░ ▒░ ▒ ░ ░      ▒ ░  ░  ▒    ░ ░  ░░ ░░   ░ ▒░    ░    
-#  ░      ░     ░   ▒   ░ ░   ░    ░   ░ ░  ▒ ░ ░ ░    ▒ ░░           ░      ░   ░ ░   ░      
-#         ░         ░  ░      ░          ░  ░          ░  ░ ░         ░  ░         ░          # 
-#               ██▓ ███▄    █  ██ ▄█▀▄▄▄     ▄▄▄█████▓ ██▓    ▄▄▄        ██████           
-#              ▓██▒ ██ ▀█   █  ██▄█▒▒████▄   ▓  ██▒ ▓▒▓██▒   ▒████▄    ▒██    ▒           
-#              ▒██▒▓██  ▀█ ██▒▓███▄░▒██  ▀█▄ ▒ ▓██░ ▒░▒██░   ▒██  ▀█▄  ░ ▓██▄             
-#              ░██░▓██▒  ▐▌██▒▓██ █▄░██▄▄▄▄██░ ▓██▓ ░ ▒██░   ░██▄▄▄▄██   ▒   ██▒          
-#              ░██░▒██░   ▓██░▒██▒ █▄▓█   ▓██▒ ▒██▒ ░ ░██████▒▓█   ▓██▒▒██████▒▒          
-#              ░▓  ░ ▒░   ▒ ▒ ▒ ▒▒ ▓▒▒▒   ▓▒█░ ▒ ░░   ░ ▒░▓  ░▒▒   ▓▒█░▒ ▒▓▒ ▒ ░          
-#               ▒ ░░ ░░   ░ ▒░░ ░▒ ▒░ ▒   ▒▒ ░   ░    ░ ░ ▒  ░ ▒   ▒▒ ░░ ░▒  ░ ░          
-#               ▒ ░   ░   ░ ░ ░ ░░ ░  ░   ▒    ░        ░ ░    ░   ▒   ░  ░  ░            
-#               ░           ░ ░  ░        ░  ░            ░  ░     ░  ░      ░            
-#         ▄████ ▓█████  ███▄    █ ▓█████  ██▀███   ▄▄▄     ▄▄▄█████▓ ▒█████   ██▀███  
-#        ██▒ ▀█▒▓█   ▀  ██ ▀█   █ ▓█   ▀ ▓██ ▒ ██▒▒████▄   ▓  ██▒ ▓▒▒██▒  ██▒▓██ ▒ ██▒
-#       ▒██░▄▄▄░▒███   ▓██  ▀█ ██▒▒███   ▓██ ░▄█ ▒▒██  ▀█▄ ▒ ▓██░ ▒░▒██░  ██▒▓██ ░▄█ ▒
-#       ░▓█  ██▓▒▓█  ▄ ▓██▒  ▐▌██▒▒▓█  ▄ ▒██▀▀█▄  ░██▄▄▄▄██░ ▓██▓ ░ ▒██   ██░▒██▀▀█▄  
-#       ░▒▓███▀▒░▒████▒▒██░   ▓██░░▒████▒░██▓ ▒██▒ ▓█   ▓██▒ ▒██▒ ░ ░ ████▓▒░░██▓ ▒██▒
-#        ░▒   ▒ ░░ ▒░ ░░ ▒░   ▒ ▒ ░░ ▒░ ░░ ▒▓ ░▒▓░ ▒▒   ▓▒█░ ▒ ░░   ░ ▒░▒░▒░ ░ ▒▓ ░▒▓░
-#         ░   ░  ░ ░  ░░ ░░   ░ ▒░ ░ ░  ░  ░▒ ░ ▒░  ▒   ▒▒ ░   ░      ░ ▒ ▒░   ░▒ ░ ▒░
-#       ░ ░   ░    ░      ░   ░ ░    ░     ░░   ░   ░   ▒    ░      ░ ░ ░ ▒    ░░   ░ 
-#             ░    ░  ░         ░    ░  ░   ░           ░  ░            ░ ░     ░     
-########################################################################################################                                                                             
-
-# STOP! 
-
-# YOU DO NOT NEED TO EDIT THE SCRIPT BELOW IN ORDER TO USE IT
-
-# RUN THIS SCRIPT FROM THE COMMAND LINE
-
-# SIMPLY TYPE CMD IN THE ADDRESS BAR OF THE FOLDER THIS SCRIPT IS LOCATED IN AND THEN PASTE
-
-# python generate_inkatlas.py 
-
-# INTO THE COMMAND LINE
-
-# DO NOT COPY THE '#' 
-
-# FOR MORE HELP: https://wiki.redmodding.org/cyberpunk-2077-modding/for-mod-creators/modding-guides/everything-else/running-python-scripts
-
-########################################################################################################  
-import os
-import sys
-import re
-import json
-try:
-    from PIL import Image
-except ImportError:
-    print('')
-    print('INKATLAS GENERATOR ERROR: PILLOW NOT INSTALLED ')
-    print('-----------------------------------------------------------------------------------------------')
-    print('')
-    print('you need to install PILLOW (Python Imaging Library) by copy pasting the following into the command line:')
-    print('')
-    print('pip install pillow')
-    print('')
-    print("for more help check the wiki: ")
-    print("https://wiki.redmodding.org/cyberpunk-2077-modding/for-mod-creators/modding-guides/everything-else/running-python-scripts")
-    print('-----------------------------------------------------------------------------------------------')
-    sys.exit(1)
-
-def main():
-    # Prompt the user for the required info
-    print('')
-    print("INKATLAS GENERATOR INPUT: ")
-    while True:
-        icon_folder = input("    Enter the path to the folder containing your individual icon PNG images: ")
-        # Check for quotes which will break the input path
-        if (icon_folder.startswith('"') or icon_folder.startswith("'")) and (icon_folder.endswith('"') or icon_folder.endswith("'")):
-            # Remove the quotes
-            icon_folder = icon_folder[1:-1]
-        if icon_folder.endswith('\\'):
-            icon_folder = icon_folder[-1]
-            
-        if not os.path.exists(icon_folder):
-            print('')
-            print('INKATLAS GENERATOR ERROR:')
-            print('-----------------------------------------------------------------------------------------------')
-            print('')
-            print('    The entered path does not exist.')
-            print('')
-            print('-----------------------------------------------------------------------------------------------')
-            print('')
-            print("INKATLAS GENERATOR INPUT: ")
-            continue
-        files_in_folder = os.listdir(icon_folder)
-
-        # Check if the icon folder contains at least one .png file
-        png_files = [file for file in files_in_folder if file.endswith('.png')]
-        if not png_files:
-            print('')
-            print('INKATLAS GENERATOR ERROR:')
-            print('-----------------------------------------------------------------------------------------------')
-            print('')
-            print('    The entered folder does not contain any PNG files.')
-            print('')
-            print('-----------------------------------------------------------------------------------------------')
-            print('')
-            print("INKATLAS GENERATOR INPUT: ")
-            continue
-
-        break  # Exit the loop if the folder contains at least one .png file
-    
-    while True:
-        output_folder = input("    Enter the path to output the raw inkatlas files for you to import in Wolvenkit: ")
-
-        # Check for quotes which will break the input path
-        if (output_folder.startswith('"') or output_folder.startswith("'")) and (output_folder.endswith('"') or output_folder.endswith("'") or output_folder.endswith("\\")):
-            # Remove the quotes
-            output_folder = output_folder[1:-1]
-
-        if output_folder.endswith("source"):
-            output_folder = os.path.join(output_folder, "raw")
-        elif "\\archive" in output_folder:
-            output_folder = output_folder.replace("\\archive", "\\raw")
-            
-        if "raw" in output_folder:
-            break  # Exit the loop if "raw" is in the output folder path
-        else:
-            print('')
-            print('INKATLAS GENERATOR ERROR: ')
-            print('-----------------------------------------------------------------------------------------------')
-            print('')
-            print('    Entered path does not seem to be within a Wolvenkit project.')
-            print('')
-            print('-----------------------------------------------------------------------------------------------')
-            print('')
-            print("INKATLAS GENERATOR INPUT: ")
-
-        
-    atlas_name = input("    Enter the name for your new inkatlas file (without extension): ")
-
-    # Load each image to get its dimensions
-    images = []
-    for png_file in png_files:
-        image_path = os.path.join(icon_folder, png_file)
-        try:
-            img = Image.open(image_path)
-            images.append({"path": image_path, "image": img, "name": os.path.splitext(png_file)[0]})
-        except Exception as e:
-            print("INKATLAS GENERATOR ERROR: ")
-            print('-----------------------------------------------------------------------------------------------')
-            print('')
-            print(f"     Error opening image {png_file}: {e}")
-            print('')
-            print('-----------------------------------------------------------------------------------------------')
-            print('')
-
-    # Calculate the maximum width of the combined image
-    max_width = 2048
-
-    # Initialize variables
-    grid = []
-    current_row = []
-    current_width = 0
-    max_height_in_row = 0
-
-    # Iterate through each image
-    for image_data in images:
-        img = image_data["image"]
-        width = img.width
-        height = img.height
-
-        # Check if adding the current image exceeds the maximum width
-        if current_width + width <= max_width:
-            # Add the image to the current row
-            current_row.append(image_data)
-            current_width += width + 1  # Add 1 pixel spacing between images
-            max_height_in_row = max(max_height_in_row, height)
-        else:
-            # Add the current row to the grid and start a new row
-            grid.append(current_row)
-            current_row = [image_data]
-            current_width = width + 1  # Add 1 pixel spacing between images
-            max_height_in_row = height
-
-    # Add the last row to the grid
-    if current_row:
-        grid.append(current_row)
-
-    # Calculate the total height of the combined image
-    total_height = sum(max(image_data["image"].height for image_data in row) for row in grid) + len(grid) - 1  # Add spacing between rows
-
-    # Ensure total height is an even number
-    total_height += total_height % 2
-
-    total_width = 0
-    for row in grid:
-        row_width = sum(image_data["image"].width + 1 for image_data in row)  # Calculate the total width of images in the current row
-        total_width = max(total_width, row_width)  # Update the total width if the current row width is greater
-
-    # Ensure total width does not exceed the maximum width
-    total_width = min(total_width, max_width)
-
-    # Ensure total width is an even number
-    total_width += total_width % 2
-
-    # Create a blank canvas to paste images onto
-    combined_image = Image.new("RGBA", (total_width, total_height), (0, 0, 0, 0))
-    raw_index = re.sub(r".*?" + re.escape("raw"), "", output_folder, 1)
-    if raw_index is not None:
-        atlas_xbm = os.path.join(raw_index, atlas_name + ".xbm")
-        atlas_xbm_1080 = atlas_xbm.replace(".xbm", "_1080.xbm")
-    # JSON data
-    data = {
-        "Header": {
-            "WolvenKitVersion": "8.13.0-nightly.2024-03-17",
-            "WKitJsonVersion": "0.0.8",
-            "GameVersion": 2120,
-            "ExportedDateTime": "2024-03-18T04:38:07.1672443Z",
-            "DataType": "CR2W",
-            "ArchiveFileName": ""
-        },
-        "Data": {
-            "Version": 195,
-            "BuildVersion": 0,
-            "RootChunk": {
-                "$type": "inkTextureAtlas",
-                "activeTexture": "StaticTexture",
-                "cookingPlatform": "PLATFORM_None",
-                "dynamicTexture": {
-                    "DepotPath": {"$type": "ResourcePath", "$storage": "uint64", "$value": "0"},
-                    "Flags": "Default"
-                },
-                "dynamicTextureSlot": {
-                    "$type": "inkDynamicTextureSlot",
-                    "parts": [],
-                    "texture": {"DepotPath": {"$type": "ResourcePath", "$storage": "uint64", "$value": "0"},
-                                "Flags": "Default"}
-                },
-                "isSingleTextureMode": 1,
-                "parts": [],
-                "slices": [],
-                "slots": {
-                    "Elements": [
-                        {
-                            "$type": "inkTextureSlot",
-                            "parts": [],
-                            "slices": [],
-                            "texture": {
-                                "DepotPath": {
-                                    "$type": "ResourcePath",
-                                    "$storage": "string",
-                                    "$value": (f"{atlas_xbm}")
-                                },
-                                "Flags": "Soft"
-                            }
-                        },
-                        {
-                            "$type": "inkTextureSlot",
-                            "parts": [],
-                            "slices": [],
-                            "texture": {
-                                "DepotPath": {
-                                    "$type": "ResourcePath",
-                                    "$storage": "string",
-                                    "$value": (f"{atlas_xbm_1080}")
-                                },
-                                "Flags": "Soft"
-                            }
-                        }
-                    ]
-                },
-                "texture": {
-                    "DepotPath": {
-                        "$type": "ResourcePath",
-                        "$storage": "string",
-                        "$value": ""
-                    },
-                    "Flags": "Default"
-                },
-                "textureResolution": "UltraHD_3840_2160"
-            },
-            "EmbeddedFiles": []
-        }
-    }
-
-    # Paste each image onto the canvas and add its data to the JSON
-    current_y = 0
-    for row in grid:
-        max_height_in_row = max(image_data["image"].height for image_data in row)
-        current_x = 0
-        for image_data in row:
-            img = image_data["image"]
-            name = image_data["name"]
-            width = img.width
-            height = img.height
-
-            top_pixel = current_y + (max_height_in_row - height) // 2
-            left_pixel = current_x
-
-            combined_image.paste(img, (left_pixel, top_pixel))
-
-            current_x += width + 1  # Add 1 pixel spacing between images
-
-            # Add image data to JSON
-            part_data = {
-                "$type": "inkTextureAtlasMapper",
-                "clippingRectInPixels": {
-                    "$type": "Rect",
-                    "bottom": top_pixel + height,
-                    "left": left_pixel,
-                    "right": left_pixel + width,
-                    "top": top_pixel
-                },
-                "clippingRectInUVCoords": {
-                    "$type": "RectF",
-                    "Bottom": (top_pixel + height) / total_height,
-                    "Left": left_pixel / total_width,
-                    "Right": (left_pixel + width) / total_width,
-                    "Top": top_pixel / total_height
-                },
-                "partName": {
-                    "$type": "CName",
-                    "$storage": "string",
-                    "$value": name
-                }
-            }
-
-            # Append part data to parts array 
-            data["Data"]["RootChunk"]["slots"]["Elements"][0]["parts"].append(part_data)
-            data["Data"]["RootChunk"]["slots"]["Elements"][1]["parts"].append(part_data)
-        current_y += max_height_in_row + 1  # Add 1 pixel spacing between rows
-        
-    # Create the output folder if it does not exist                             
-    if not os.path.exists(output_folder):
-        os.makedirs(output_folder)
-
-    # Set the output file name
-    output_file = os.path.join(output_folder, atlas_name + ".inkatlas.json")
-
-    # Write everything to the .inkatlas.json
-    with open(output_file, "w") as json_file:
-        json.dump(data, json_file, indent=2)
-
-    print(f"Data has been saved to {output_file}")
-
-    # Save the combined image
-    combined_image_path = os.path.join(output_folder, atlas_name + ".png")
-    combined_image.save(combined_image_path)
-    print(f"Combined image has been saved to {combined_image_path}")
-    # Save the resized image with "_1080" suffix
-    resized_image = combined_image.resize((total_width // 2, total_height // 2),resample=None, box=None, reducing_gap=None)
-    combined_image_path_1080 = combined_image_path.replace(".png", "_1080.png")
-    resized_image.save(combined_image_path_1080)    
-    print(f"Combined image has been saved to {combined_image_path_1080}")
-
-if __name__ == "__main__":
-    main()
